+++ python/marvin/__init__.py
@@ -14,26 +14,20 @@ import contextlib
 import yaml
 import six
 from collections import OrderedDict
+from astropy.wcs import FITSFixedWarning
 
 # Set the Marvin version
 __version__ = '2.3.0dev'
 
 # Does this so that the implicit module definitions in extern can happen.
+# time - 483 ms
 from marvin import extern
-
+from marvin.core.exceptions import MarvinUserWarning, MarvinError
 from brain.utils.general.general import getDbMachine, merge
 from brain import bconfig
 from brain.core.core import URLMapDict
 from brain.core.exceptions import BrainError
 from brain.core.logger import initLog
-
-from astropy.wcs import FITSFixedWarning
-
-# Set the Marvin version
-__version__ = '2.2.6dev'
-
-from marvin.core.exceptions import MarvinUserWarning, MarvinError
-
 
 # Defines log dir.
 if 'MARVIN_LOGS_DIR' in os.environ:
@@ -58,6 +52,7 @@ warnings.filterwarnings('ignore', category=DeprecationWarning)
 warnings.filterwarnings('once', category=DeprecationWarning, module='marvin')
 
 
+# up to here - time: 1 second
 class MarvinConfig(object):
     ''' Global Marvin Configuration
 
@@ -302,6 +297,14 @@ class MarvinConfig(object):
         bconfig.traceback = value
 
     @property
+    def compression(self):
+        return bconfig.compression
+
+    @compression.setter
+    def compression(self, value):
+        bconfig.compression = value
+
+    @property
     def token(self):
         return bconfig.token
 
@@ -317,7 +320,7 @@ class MarvinConfig(object):
 
         if self._urlmap is None or (isinstance(self._urlmap, dict) and len(self._urlmap) == 0):
             try:
-                response = Interaction('api/general/getroutemap', request_type='get', auth='netrc')
+                response = Interaction('/marvin/api/general/getroutemap', request_type='get', auth='netrc')
             except Exception as e:
                 warnings.warn('Cannot retrieve URLMap. Remote functionality will not work: {0}'.format(e),
                               MarvinUserWarning)
@@ -373,8 +376,7 @@ class MarvinConfig(object):
                    'MPL-5': ('v2_0_1', '2.0.2'),
                    'MPL-4': ('v1_5_1', '1.1.1')}
 
-        drdict = {'DR15': ('v2_4_3', '2.2.1'),
-                  'DR14': ('v2_1_2', '2.0.2')}
+        drdict = {'DR15': ('v2_4_3', '2.2.1')}
 
         # set the allowed releases based on access
         self._allowed_releases = {}
@@ -526,21 +528,39 @@ class MarvinConfig(object):
         '''
         assert sasmode in ['utah', 'local'], 'SAS mode can only be utah or local'
         base = base if base else os.environ.get('MARVIN_BASE', 'marvin')
+        test_domain = 'https://lore.sdss.utah.edu/'
         if sasmode == 'local':
             if ngrokid:
-                self.sasurl = 'http://{0}.ngrok.io/{1}/'.format(ngrokid, base)
+                self.sasurl = 'http://{0}.ngrok.io/'.format(ngrokid)
             else:
-                self.sasurl = 'http://localhost:{0}/{1}/'.format(port, base)
+                self.sasurl = 'http://localhost:{0}/'.format(port)
         elif sasmode == 'utah':
-            marvin_base = 'test/{0}/'.format(base) if test else '{0}/'.format(base)
+            # sort out the domain name
             if public:
-                base_url = re.sub(r'(dr[0-9]{1,2})', self._release.lower(), bconfig.public_api_url)
-                public_api = os.path.join(base_url, marvin_base)
-                self.sasurl = public_api
-                self._authtype = None
+                domain = test_domain if test else bconfig.public_api_url
+                if test:
+                    domain = '{0}{1}'.format(domain, 'public/')
+                else:
+                    domain = re.sub(r'(dr[0-9]{1,2})', self._release.lower(), domain)
             else:
-                self.sasurl = os.path.join(bconfig.collab_api_url, marvin_base)
-                self._authtype = 'token'
+                domain = test_domain if test else bconfig.collab_api_url
+
+            url = os.path.join(domain)
+            self.sasurl = url
+
+            # get a new urlmap (need to do for vetting)
+            self._urlmap = None
+            #__ = self.urlmap
+
+            # marvin_base = 'test/{0}/'.format(base) if test else '{0}/'.format(base)
+            # if public:
+            #     base_url = re.sub(r'(dr[0-9]{1,2})', self._release.lower(), bconfig.public_api_url)
+            #     public_api = os.path.join(base_url, marvin_base)
+            #     self.sasurl = public_api
+            #     self._authtype = None
+            # else:
+            #     self.sasurl = os.path.join(bconfig.collab_api_url, marvin_base)
+            #     self._authtype = 'token'
 
     def forceDbOff(self):
         ''' Force the database to be turned off '''
@@ -689,10 +709,12 @@ class MarvinConfig(object):
 
 
 config = MarvinConfig()
+# up to here - time: 1.6 seconds
 
-# Inits the Database session and ModelClasses
-from marvin.db.marvindb import MarvinDB
-marvindb = MarvinDB(dbtype=config.db)
+# Inits the Database session and ModelClasses (time: 1.8 seconds)
+if config.db:
+    from marvin.db.marvindb import MarvinDB
+    marvindb = MarvinDB(dbtype=config.db, log=log, allowed_releases=config._allowed_releases.keys())
 
 # Init MARVIN_DIR
 marvindir = os.environ.get('MARVIN_DIR', None)