+++ python/marvin/api/__init__.py
@@ -67,10 +67,14 @@ params = {'query': {'searchfilter': fields.String(allow_none=True),
                     'sort': fields.String(allow_none=True),
                     'order': fields.String(missing='asc', validate=validate.OneOf(['asc', 'desc'])),
                     'rettype': fields.String(allow_none=True, validate=validate.OneOf(['cube', 'spaxel', 'maps', 'rss', 'modelcube'])),
-                    'params': fields.DelimitedList(fields.String(), allow_none=True),
+                    'returnparams': fields.DelimitedList(fields.String(), allow_none=True),
+                    'defaults': fields.DelimitedList(fields.String(), allow_none=True),
+                    'targets': fields.DelimitedList(fields.String(), allow_none=True),
+                    'quality': fields.DelimitedList(fields.String(), allow_none=True),
                     'return_all': fields.Boolean(allow_none=True),
                     'format_type': fields.String(allow_none=True, validate=validate.OneOf(['list', 'listdict', 'dictlist'])),
-                    'caching': fields.Boolean(allow_none=True)
+                    'caching': fields.Boolean(allow_none=True),
+                    'query_type': fields.String(allow_none=True, validate=validate.OneOf(['raw', 'core', 'orm']))
                     },
           'search': {'searchbox': fields.String(required=True),
                      'parambox': fields.DelimitedList(fields.String(), allow_none=True)