+++ python/marvin/tests/tools/test_results.py
@@ -27,21 +27,21 @@ import json
 
 
 myplateifu = '8485-1901'
-cols = ['cube.mangaid', 'cube.plate', 'cube.plateifu', 'ifu.name', 'nsa.z']
-remotecols = [u'mangaid', u'plate', u'plateifu', u'ifu_name', u'z']
+cols = ['cube.mangaid', 'cube.plateifu', 'nsa.z']
+remotecols = [u'mangaid', u'plateifu', u'z']
 
 
 @pytest.fixture()
 def limits(results):
-    data = results.expdata['queries'][results.searchfilter]
+    data = results.expdata['queries'][results.search_filter]
     count = 10 if data['count'] >= 10 else data['count']
     return (10, count)
 
 
 @pytest.fixture()
 def results(query, request):
-    searchfilter = request.param if hasattr(request, 'param') else 'nsa.z < 0.1 and cube.plate==8485'
-    q = Query(searchfilter=searchfilter, mode=query.mode, limit=10, release=query._release)
+    searchfilter = request.param if hasattr(request, 'param') else 'nsa.z < 0.1'
+    q = Query(search_filter=searchfilter, mode=query.mode, limit=10, release=query.release)
     r = q.run()
     r.expdata = query.expdata
     yield r
@@ -57,10 +57,10 @@ def columns(results):
 class TestResultsColumns(object):
 
     def test_check_cols(self, results, columns):
-        assert results.columns.full == columns.full
-        assert results.columns.remote == columns.remote
+        assert set(results.columns.full) == set(columns.full)
+        assert set(results.columns.remote) == set(columns.remote)
 
-    @pytest.mark.parametrize('col, errmsg', [('plate', 'plate is too ambiguous.  Did you mean one of')])
+    @pytest.mark.parametrize('col, errmsg', [('cube', 'cube is too ambiguous.  Did you mean one of')])
     def test_fail(self, results, col, errmsg):
         with pytest.raises(KeyError) as cm:
             col in results.columns
@@ -94,29 +94,38 @@ class TestResultsColumns(object):
 class TestMarvinTuple(object):
 
     def test_create(self, results):
-        data = results.results[0].__dict__
+        data = results.results[0]._asdict()
         mt = marvintuple('ResultRow', data.keys())
         assert mt is not None
         assert hasattr(mt, 'mangaid')
         assert hasattr(mt, 'plateifu')
 
         row = mt(**data)
-        assert row.plate == data['plate']
         assert row.mangaid == data['mangaid']
         assert row.plateifu == data['plateifu']
 
-    def test_add(self, results):
-        data = results.results[0].__dict__
-        mt = marvintuple('ResultRow', 'mangaid, plate, plateifu, ifu_name')
-        mt1 = marvintuple('ResultRow', 'z')
+    @pytest.mark.parametrize('params, msg',
+                             [('plateifu, z', None),
+                              ('z', 'All rows must have a plateifu column to be able to add')],
+                             ids=['pass', 'fail'])
+    def test_add(self, results, params, msg):
+        data = results.results[0]._asdict()
+        mt = marvintuple('ResultRow', 'mangaid, plateifu')
+        mt1 = marvintuple('ResultRow', params)
+        cols = [c.strip() for c in params.split(',')]
+        row = mt(**{k: v for k, v in data.items() if k in ['mangaid', 'plateifu']})
+        row1 = mt1(**{k: v for k, v in data.items() if k in cols})
 
-        row = mt(**{k: v for k, v in data.items() if k in ['mangaid', 'plate', 'plateifu', 'ifu_name']})
-        row1 = mt1(**{k: v for k, v in data.items()})
-        new_row = row + row1
-        assert new_row is not None
-        cols = ['mangaid', 'plate', 'plateifu', 'z']
-        assert set(cols).issubset(set(new_row.__dict__.keys()))
-        assert row.__dict__ <= new_row.__dict__
+        if msg:
+            with pytest.raises(AssertionError) as cm:
+                new_row = row + row1
+            assert msg in str(cm)
+        else:
+            new_row = row + row1
+            assert new_row is not None
+            cols = ['mangaid', 'plateifu'] + cols
+            assert set(cols).issubset(set(new_row._asdict().keys()))
+            assert all(item in new_row._asdict().items() for item in row._asdict().items())
 
 
 class TestResultSet(object):
@@ -181,7 +190,7 @@ class TestResultsGetParams(object):
 
     @pytest.mark.parametrize('col',
                              [(c) for c in cols],
-                             ids=['cube.mangaid', 'cube.plate', 'cube.plateifu', 'ifu.name', 'nsa.z'])
+                             ids=['cube.mangaid', 'cube.plateifu', 'nsa.z'])
     def test_get_list(self, results, col):
         assert col in results.columns
         obj = results.getListOf(col)
@@ -191,17 +200,17 @@ class TestResultsGetParams(object):
         assert isinstance(json_obj, six.string_types)
 
     @pytest.mark.parametrize('results',
-                             [('nsa.z < 0.1 and haflux > 25'),
+                             [('nsa.z < 0.1 and emline_gflux_ha_6564 > 25'),
                               ('nsa.z < 0.1'),
-                              ('haflux > 25')], indirect=True)
+                              ('emline_gflux_ha_6564 > 25')], indirect=True)
     def test_get_list_all(self, results):
-        q = Query(searchfilter=results.searchfilter, mode=results.mode, limit=1,
-                  release=results._release, returnparams=results.returnparams)
+        q = Query(search_filter=results.search_filter, mode=results.mode, limit=1,
+                  release=results.release, return_params=results.return_params)
         r = q.run(start=0, end=1)
         assert r.count == 1
         mangaids = r.getListOf('mangaid', return_all=True)
         assert len(mangaids) == r.totalcount
-        assert len(mangaids) == results.expdata['queries'][results.searchfilter]['count']
+        assert len(mangaids) == results.expdata['queries'][results.search_filter]['count']
 
     @pytest.mark.parametrize('ftype', [('dictlist'), ('listdict')])
     @pytest.mark.parametrize('name', [(None), ('mangaid'), ('z')], ids=['noname', 'mangaid', 'z'])
@@ -237,7 +246,7 @@ class TestResultsSort(object):
     def test_sort(self, results, limits):
         results.sort('z')
         limit, count = limits
-        data = results.expdata['queries'][results.searchfilter]['sorted']
+        data = results.expdata['queries'][results.search_filter]['sorted']
         assert tuple(data['1']) == results.results[0]
         assert tuple(data[str(count)]) == results.results[count - 1]
 
@@ -250,7 +259,7 @@ class TestResultsPaging(object):
             pytest.skip('skipping now due to weird issue with local results not same as remote results')
         results.sort('z')
         limit, count = limits
-        data = results.expdata['queries'][results.searchfilter]
+        data = results.expdata['queries'][results.search_filter]
         assert results.totalcount == data['count']
         assert results.count == count
         assert len(results.results) == count
@@ -268,7 +277,7 @@ class TestResultsPaging(object):
         limit, count = limits
         results.sort('z')
         results.getNext(chunk=chunk)
-        data = results.expdata['queries'][results.searchfilter]['sorted']
+        data = results.expdata['queries'][results.search_filter]['sorted']
         if results.count == results.totalcount:
             assert results.results[0] == tuple(data['1'])
             assert len(results.results) == count
@@ -290,7 +299,7 @@ class TestResultsPaging(object):
         results.sort('z')
         results.getSubset(index, limit=chunk)
         results.getPrevious(chunk=chunk)
-        data = results.expdata['queries'][results.searchfilter]['sorted']
+        data = results.expdata['queries'][results.search_filter]['sorted']
         if results.count == results.totalcount:
             assert results.results[0] == tuple(data['1'])
             assert len(results.results) == count
@@ -313,7 +322,7 @@ class TestResultsPaging(object):
         limit, count = limits
         results.sort('z')
         results.getSubset(index, limit=chunk)
-        data = results.expdata['queries'][results.searchfilter]['sorted']
+        data = results.expdata['queries'][results.search_filter]['sorted']
         if results.count == results.totalcount:
             assert results.results[0] == tuple(data['1'])
             assert len(results.results) == count
@@ -340,7 +349,7 @@ class TestResultsPaging(object):
         res = results.getSubset(0, limit=1)
         assert results.count == 1
         results.loop(chunk=500)
-        assert results.count == results.expdata['queries'][results.searchfilter]['count']
+        assert results.count == results.expdata['queries'][results.search_filter]['count']
         assert results.count == results.totalcount
 
     @pytest.mark.parametrize('results', [('nsa.z < 0.1')], indirect=True)
@@ -361,7 +370,7 @@ class TestResultsPickling(object):
         path = results.save(str(file), overwrite=True)
         assert file.check() is True
         r = Results.restore(str(file))
-        assert r.searchfilter == results.searchfilter
+        assert r.search_filter == results.search_filter
 
 
 class TestResultsConvertTool(object):
@@ -370,11 +379,11 @@ class TestResultsConvertTool(object):
     @pytest.mark.parametrize('objtype, tool',
                              [('cube', Cube), ('maps', Maps), ('spaxel', Spaxel),
                               ('modelcube', ModelCube)])
-    def test_convert_success(self, results, objtype, tool, exporigin):
+    def test_convert_success(self, results, objtype, tool):
         if config.release == 'MPL-4' and objtype == 'modelcube':
             pytest.skip('no modelcubes in mpl-4')
 
-        results.convertToTool(objtype, limit=1)
+        results.convertToTool(objtype, limit=1, mode=results.mode)
         assert results.objects is not None
         assert isinstance(results.objects, list) is True
         assert isinstance(results.objects[0], tool) is True
@@ -382,7 +391,7 @@ class TestResultsConvertTool(object):
             assert results.mode == results.objects[0].mode
 
     @pytest.mark.parametrize('objtype, error, errmsg',
-                             [('modelcube', MarvinError, "ModelCube requires at least dapver='2.0.2'"),
+                             [('modelcube', AssertionError, "ModelCubes require a release of MPL-5 and up"),
                               ('spaxel', AssertionError, 'Parameters must include spaxelprop.x and y in order to convert to Marvin Spaxel')],
                              ids=['mcminrelease', 'nospaxinfo'])
     def test_convert_failures(self, results, objtype, error, errmsg):